chat-backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
pydantic
passlib[bcrypt]
python-jose[cryptography]
typing
python-multipart
sqlalchemy
databases[postgresql]
